requirements.txt
absl-py==2.1.0
einops==0.8.0
lpips==0.1.4
mamba-ssm==2.2.2
markdown==3.7
matplotlib==3.7.5
mkl==2021.4.0
numpy==1.24.1
pandas==2.2.2
pillow==10.4.0
protobuf==5.29.3
psutil==7.0.0
pytorch==2.1.1
pytz==2025.1
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-learn==1.5.1
scipy==1.14.1
seaborn==0.13.2
setuptools==75.8.0
tensorboard==2.19.0
tensorboardx==2.6.2.2
threadpoolctl==3.5.0
timm==1.0.9
tokenizers==0.21.0
torchaudio==2.1.1
torchvision==0.16.1
tqdm==4.66.4
transformers==4.48.3
triton==2.0.0
typing_extensions==4.12.2
urllib3==2.3.0
wheel==0.45.1
